pyproject.toml
[tool.poetry]
name = "alpaca-testing"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alpaca-py = "^0.10.0"
python-dotenv = "^1.0.0"
black = "^23.9.1"
flask = "^2.3.3"
werkzeug = "^2.3.7"
functions-framework = "^3.4.0"
google-cloud-logging = "^3.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
